fix scripts

+++ Actions with orders, charges and invoices/recalculate_charge_with_actual_price.rb
@@ -5,8 +5,10 @@ def recalculate_charge_with_actual_price(charge_id)
     charge.price = new_price
     charge.recalculate_amount!
     charge.save
-    ResellerCharge.where(charge_id: charge.id).delete_all
-    ResellerChargesBuilder.new(charge).call
+    if charge.status == 'closed'
+      ResellerCharge.where(charge_id: charge.id).delete_all
+      ResellerChargesBuilder.new(charge).call
+    end
     order = charge.order
     OrderDetailsGenerator.new(order).generate!
     order.recalculate_total!
